requirements.txt
asgiref==3.7.2
astroid==2.15.6
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
decorator==4.4.2
dill==0.3.7
Django==4.2.4
django-ckeditor==6.7.0
django-js-asset==2.1.0
geographiclib==2.0
geopy==2.4.0
gpxpy==1.5.0
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mod-wsgi==4.9.4
moviepy==1.0.3
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
platformdirs==3.10.0
proglog==0.1.10
pylint==2.17.5
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
requests==2.31.0
sqlparse==0.4.4
tomlkit==0.12.1
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
